--- backend/src/main/java/com/woochacha/backend/domain/admin/dto/CarInspectionRequestDto.java
@@ -1,13 +1,0 @@
-package com.woochacha.backend.domain.admin.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-
-@Getter
-@Builder
-public class CarInspectionRequestDto {
-    private String carNum;
-    private int distance;
-}
